Move HIP-RA-X to separate WIP hip_ra_x module and revert hip_ra module to equivalent of https://github.com/NREL/GEOPHIRES-X/commit/3d5d75477ba93df399f66646d433ff56a8e6bbe5 for now.
