+++ src/hip_ra_x/.gitignore
@@ -1,0 +1,2 @@
+*.log
+*.out
