+++ src/hip_ra_x/logging.conf
@@ -1,0 +1,31 @@
+[loggers]
+keys=root
+
+[handlers]
+keys=consoleHandler, fileHandler
+
+[formatters]
+keys=consoleFormatter, fileFormatter
+
+[logger_root]
+level=WARN
+handlers=consoleHandler, fileHandler
+
+[handler_consoleHandler]
+class=logging.StreamHandler
+level=WARN
+formatter=consoleFormatter
+args=(sys.stdout,)
+
+[handler_fileHandler]
+class = logging.FileHandler
+kwargs = {"filename": "all_messages_conf.log"}
+level = INFO
+formatter = fileFormatter
+
+[formatter_consoleFormatter]
+format=%(asctime)s - %(name)s - %(levelname)s - %(message)s
+
+[formatter_fileFormatter]
+format = %(asctime)s : %(levelname)s : %(module)s : %(funcName)s : %(lineno)d : (Process Details : (%(process)d, %(processName)s), Thread Details : (%(thread)d, %(threadName)s)): %(message)s
+datefmt = %d-%m-%Y %I:%M:%S
