+++ tests/examples/HIPexample1.out
@@ -7,10 +7,10 @@
       Reservoir Temperature:       250.00 degC
       Reservoir Volume:             13.75 km**3
       Stored Heat:               7.42e+15 kJ
-      Fluid Produced:            1.10e+12 kilogram
+      Fluid Produced:            1.09e+12 kilogram
       Enthalpy:                    181.43 kJ/kg
-      Wellhead Heat:             8.27e+14 kJ
-      Recovery Factor:              11.15 %
-      Available Heat:            1.47e+13 kJ
-      Producible Heat:           5.86e+12 kJ
-      Producible Electricity:      185.85 MW
+      Wellhead Heat:             8.19e+14 kJ
+      Recovery Factor:              11.04 %
+      Available Heat:            1.44e+13 kJ
+      Producible Heat:           5.75e+12 kJ
+      Producible Electricity:      182.47 MW
