+++ tests/test_HIP_RA.py
@@ -1,27 +1,8 @@
-import logging
-import os
-import re
-import sys
-import tempfile
-import unittest
-import uuid
 from pathlib import Path
 
-from geophires_x.Parameter import OutputParameter
-from geophires_x.Parameter import ParameterEntry
-from geophires_x.Parameter import floatParameter
-from geophires_x.Parameter import intParameter
-from geophires_x.Units import EnthalpyUnit
-from geophires_x.Units import HeatUnit
-from geophires_x.Units import PercentUnit
-from geophires_x.Units import PowerUnit
-from geophires_x.Units import TemperatureUnit
-from geophires_x.Units import Units
-from geophires_x.Units import VolumeUnit
 from hip_ra import HipRaClient
 from hip_ra import HipRaInputParameters
 from hip_ra import HipRaResult
-from hip_ra.HIP_RA import HIP_RA
 from tests.base_test_case import BaseTestCase
 
 
@@ -110,327 +91,3 @@ class HIP_RATestCase(BaseTestCase):
                 'Producible Electricity/Unit Area (fluid)': {'value': 32.01, 'unit': 'MW/km**2'},
             },
         )
-
-    def test_calculate_reservoir_volume(self):
-        """Calculates the volume of the reservoir"""
-
-        hip_ra = self._new_hip_ra_test_instance()
-        hip_ra.Calculate()
-        assert hip_ra.reservoir_volume.value == hip_ra.reservoir_area.value * hip_ra.reservoir_thickness.value
-
-    def test_standard_outputs(self):
-        """Prints the standard outputs to the output file"""
-
-        hip_ra = self._new_hip_ra_test_instance()
-        hip_ra.PrintOutputs()
-
-        # Assert that the output file is created
-        # ruff: noqa: PTH110
-        assert os.path.exists('HIP.out')
-
-        # Assert that the output file is not empty
-        # ruff: noqa: PTH202
-        assert os.path.getsize('HIP.out') > 0
-
-        # Clean up the output file
-        # ruff: noqa: PTH107
-        os.remove('HIP.out')
-
-    def test_converts_units_back(self):
-        """Converts Units back to PreferredUnits, if required"""
-
-        hip_ra = self._new_hip_ra_test_instance()
-        hip_ra.PrintOutputs()
-
-        # Assert that the units of all parameters in ParameterDict are converted back to PreferredUnits
-        for key in hip_ra.ParameterDict:
-            param = hip_ra.ParameterDict[key]
-            assert param.CurrentUnits == param.PreferredUnits
-
-    def test_updates_output_parameter_units(self):
-        hip_ra = self._new_hip_ra_test_instance()
-        hip_ra.PrintOutputs()
-        # Assert that the units of all parameters in OutputParameterDict are updated to the user-specified units
-        for key in hip_ra.OutputParameterDict:
-            param: OutputParameter = hip_ra.OutputParameterDict[key]
-            assert param.CurrentUnits == param.PreferredUnits
-
-    @unittest.skip('FIXME WIP')
-    def test_aligns_space_between_value_and_units(self):
-        """
-        Assert that the space between value and units is aligned to the same column for each line in the output file
-        """
-
-        hip_ra = self._new_hip_ra_test_instance()
-        hip_ra.PrintOutputs()
-
-        with open('HIP.out') as f:
-            content = f.readlines()
-            for line in content:
-                assert line.count(' ') == 3
-
-    def test_raises_permission_error(self):
-        """Raises a PermissionError if there is no permission to write to the output file"""
-
-        hip_ra = self._new_hip_ra_test_instance()
-        # Create a read-only file
-        Path.chmod('HIP.out', 0o444)
-        with self.assertRaises(PermissionError):
-            hip_ra.PrintOutputs()
-        # Restore file permissions
-        Path.chmod('HIP.out', 0o644)
-
-    def test_handles_converting_output_units(self):
-        """Handles converting output units for all classes of units (TEMPERATURE, DENSITY, etc.)"""
-
-        hip_ra: HIP_RA = self._new_hip_ra_test_instance()
-        hip_ra.PrintOutputs()
-        # Assert that the units of all parameters in OutputParameterDict are converted to the user-specified units
-        for key in hip_ra.OutputParameterDict:
-            param = hip_ra.OutputParameterDict[key]
-            assert param.CurrentUnits == param.PreferredUnits
-
-    def test_handles_rendering_float_parameters_in_scientific_notation(self):
-        """
-        Assert that the float parameters in scientific notation are rendered correctly in the output file
-        """
-        client = HipRaClient()
-        result: HipRaResult = client.get_hip_ra_result(
-            HipRaInputParameters(self._get_test_file_path('examples/HIPexample1.txt'))
-        )
-        with open(result.output_file_path) as f:
-            content = f.read()
-            assert len(re.compile(r'[0-9]e\+[0-9]+\s').findall(content)) > 0
-
-    def test_read_all_parameters_no_effect_with_no_input_file(self):
-        def read_params(hip_ra):
-            hip_ra.read_parameters()
-
-        hip_ra: HIP_RA = self._new_hip_ra_test_instance(pre_re_stash_runner=read_params)
-
-        # Assert that all the parameters have been read in and updated
-        assert hip_ra.reservoir_temperature.value == 150.0
-        assert hip_ra.rejection_temperature.value == 25.0
-        assert hip_ra.reservoir_porosity.value == 18.0
-        assert hip_ra.reservoir_area.value == 81.0
-        assert hip_ra.reservoir_thickness.value == 0.286
-
-    def test_update_changed_parameters(self):
-        """updates any of these parameter values that have been changed by the user"""
-
-        def set_and_read_input_params(hip_ra):
-            # Set some input parameters to be changed by the user
-            hip_ra.InputParameters = {
-                param.Name: param
-                for param in [
-                    ParameterEntry(Name='Reservoir Temperature', sValue='200'),
-                    ParameterEntry(Name='Reservoir Porosity', sValue='25.0'),
-                ]
-            }
-
-            hip_ra.read_parameters()
-
-        hip_ra: HIP_RA = self._new_hip_ra_test_instance(pre_re_stash_runner=set_and_read_input_params)
-
-        # Assert that the changed parameters have been updated
-        assert hip_ra.reservoir_temperature.value == 200.0
-        assert hip_ra.reservoir_porosity.value == 25.0
-
-    def test_handle_special_cases(self):
-        def set_and_read_input_params(hip_ra):
-            # Set some input parameters to trigger special cases
-            hip_ra.InputParameters = {
-                param.Name: param
-                for param in [
-                    ParameterEntry(Name='Reservoir Porosity', sValue='30.0'),
-                    ParameterEntry(Name='Rejection Temperature', sValue='50.0'),
-                    ParameterEntry(Name='Density Of Water', sValue='-1'),
-                    ParameterEntry(Name='Heat Capacity Of Water', sValue='-1'),
-                    ParameterEntry(Name='Recoverable Heat', sValue='-1'),
-                ]
-            }
-
-            hip_ra.read_parameters()
-
-        hip_ra = self._new_hip_ra_test_instance(pre_re_stash_runner=set_and_read_input_params)
-
-        assert hip_ra.rejection_temperature.value == 50
-        assert hip_ra.fluid_density.value == -1
-        assert hip_ra.fluid_heat_capacity.value == -1
-        assert hip_ra.fluid_recoverable_heat.value == -1
-
-    def test_set_current_units_preferred_units_match(self):
-        """sets the CurrentUnits of a parameter to its PreferredUnits if they match"""
-
-        def set_and_read_input_params(hip_ra):
-            # Set some input parameters with matching PreferredUnits
-            hip_ra.InputParameters = {
-                param.Name: param
-                for param in [
-                    ParameterEntry(Name='Reservoir Temperature', sValue='150.0'),
-                    ParameterEntry(Name='Formation Porosity', sValue='18.0'),
-                ]
-            }
-
-            hip_ra.read_parameters()
-
-        hip_ra = self._new_hip_ra_test_instance(pre_re_stash_runner=set_and_read_input_params)
-
-        # Assert that the CurrentUnits have been set to PreferredUnits
-        assert hip_ra.reservoir_temperature.CurrentUnits == TemperatureUnit.CELSIUS
-        assert hip_ra.reservoir_porosity.CurrentUnits == PercentUnit.PERCENT
-
-    def test_set_current_units_preferred_units_do_not_match(self):
-        """sets the CurrentUnits of a parameter to the units provided by the user if they don't match"""
-
-        def set_and_read_input_params(hip_ra: HIP_RA):
-            # Set some input parameters with non-matching PreferredUnits
-            hip_ra.InputParameters = {
-                param.Name: param
-                for param in [
-                    # TODO Pint conversion would treat 'F' as farad, hence 'degF' here. However, it is possible to
-                    #  configure Pint to treat 'F' as Fahrenheit instead, which should be done since users could make
-                    #  this mistake.
-                    ParameterEntry(Name='Reservoir Temperature', sValue='150.0 degF'),
-                    ParameterEntry(Name='Formation Porosity', sValue='18.0 %'),
-                ]
-            }
-
-            hip_ra.read_parameters()
-
-        hip_ra: HIP_RA = self._new_hip_ra_test_instance(pre_re_stash_runner=set_and_read_input_params)
-
-        # Assert that the CurrentUnits have been set to the units provided by the user
-        assert hip_ra.reservoir_temperature.CurrentUnits == TemperatureUnit.FAHRENHEIT
-        assert hip_ra.reservoir_porosity.CurrentUnits == PercentUnit.PERCENT
-
-    def test_convert_units_of_output_parameters(self):
-        """The class converts units of the output parameters if specified in the input file."""
-
-        hip_ra: HIP_RA = self._new_hip_ra_test_instance()
-        hip_ra.OutputParameterDict['Reservoir Volume (reservoir)'].PreferredUnits = VolumeUnit.METERS3
-        hip_ra.OutputParameterDict['Stored Heat (reservoir)'].PreferredUnits = HeatUnit.J
-
-        # FIXME WIP
-        # hip_ra.OutputParameterDict['Fluid Produced'].PreferredUnits = MassUnit.GRAM
-
-        hip_ra.OutputParameterDict['Enthalpy (reservoir)'].PreferredUnits = EnthalpyUnit.KJPERKG
-        hip_ra.OutputParameterDict['Wellhead Heat (reservoir)'].PreferredUnits = HeatUnit.J
-        hip_ra.OutputParameterDict['Recovery Factor (reservoir)'].PreferredUnits = PercentUnit.PERCENT
-        hip_ra.OutputParameterDict['Available Heat (reservoir)'].PreferredUnits = HeatUnit.J
-        hip_ra.OutputParameterDict['Producible Heat (reservoir)'].PreferredUnits = HeatUnit.J
-        hip_ra.OutputParameterDict['Producible Electricity (reservoir)'].PreferredUnits = PowerUnit.W
-
-        hip_ra.PrintOutputs()
-
-        assert hip_ra.OutputParameterDict['Reservoir Volume (reservoir)'].CurrentUnits == VolumeUnit.KILOMETERS3
-        assert hip_ra.OutputParameterDict['Stored Heat (reservoir)'].CurrentUnits == HeatUnit.KJ
-
-        # FIXME WIP
-        # assert hip_ra.OutputParameterDict['Fluid Produced'].CurrentUnits == MassUnit.KILOGRAM
-
-        assert hip_ra.OutputParameterDict['Enthalpy (reservoir)'].CurrentUnits == EnthalpyUnit.KJPERKG
-        assert hip_ra.OutputParameterDict['Wellhead Heat (reservoir)'].CurrentUnits == HeatUnit.KJ
-        assert hip_ra.OutputParameterDict['Recovery Factor (reservoir)'].CurrentUnits == PercentUnit.PERCENT
-        assert hip_ra.OutputParameterDict['Available Heat (reservoir)'].CurrentUnits == HeatUnit.KJ
-        assert hip_ra.OutputParameterDict['Producible Heat (reservoir)'].CurrentUnits == HeatUnit.KJ
-        assert hip_ra.OutputParameterDict['Producible Electricity (reservoir)'].CurrentUnits == PowerUnit.MW
-
-    def test_initialization_with_default_parameters(self):
-        hip_ra: HIP_RA = self._new_hip_ra_test_instance()
-
-        assert isinstance(hip_ra.reservoir_temperature, floatParameter)
-        assert isinstance(hip_ra.rejection_temperature, floatParameter)
-        assert isinstance(hip_ra.reservoir_porosity, floatParameter)
-        assert isinstance(hip_ra.reservoir_area, floatParameter)
-        assert isinstance(hip_ra.reservoir_thickness, floatParameter)
-        assert isinstance(hip_ra.rock_heat_capacity, floatParameter)
-        assert isinstance(hip_ra.fluid_heat_capacity, floatParameter)
-        assert isinstance(hip_ra.rock_heat_capacity, floatParameter)
-        assert isinstance(hip_ra.fluid_density, floatParameter)
-        assert isinstance(hip_ra.rock_density, floatParameter)
-        assert isinstance(hip_ra.fluid_recoverable_heat, floatParameter)
-
-        # TODO should these be initialized?
-        # assert isinstance(hip_ra.rejection_temperature_k, floatParameter)
-        # assert isinstance(hip_ra.rejection_entropy, floatParameter)
-        # assert isinstance(hip_ra.rejection_enthalpy, floatParameter)
-
-        assert isinstance(hip_ra.reservoir_life_cycle, intParameter)
-
-        assert isinstance(hip_ra.volume_fluid, OutputParameter)
-        assert isinstance(hip_ra.volume_rock, OutputParameter)
-        assert isinstance(hip_ra.reservoir_volume, OutputParameter)
-        assert isinstance(hip_ra.reservoir_stored_heat, OutputParameter)
-        assert isinstance(hip_ra.reservoir_mass, OutputParameter)
-        assert isinstance(hip_ra.reservoir_enthalpy, OutputParameter)
-        assert isinstance(hip_ra.wellhead_heat, OutputParameter)
-        assert isinstance(hip_ra.reservoir_recovery_factor, OutputParameter)
-        assert isinstance(hip_ra.reservoir_available_heat, OutputParameter)
-        assert isinstance(hip_ra.reservoir_producible_heat, OutputParameter)
-        assert isinstance(hip_ra.reservoir_producible_electricity, OutputParameter)
-
-        assert hip_ra.reservoir_thickness.value == 0.286
-
-        assert hip_ra.reservoir_temperature.Name == 'Reservoir Temperature'
-        assert hip_ra.reservoir_temperature.value == 150.0
-        assert hip_ra.reservoir_temperature.Min == 50
-        assert hip_ra.reservoir_temperature.Max == 1000
-        assert hip_ra.reservoir_temperature.UnitType == Units.TEMPERATURE
-        assert hip_ra.reservoir_temperature.PreferredUnits == TemperatureUnit.CELSIUS
-        assert hip_ra.reservoir_temperature.CurrentUnits == TemperatureUnit.CELSIUS
-        assert hip_ra.reservoir_temperature.Required is True
-        assert hip_ra.reservoir_temperature.ErrMessage == 'assume default reservoir temperature (150 deg-C)'
-        assert hip_ra.reservoir_temperature.ToolTipText == 'Reservoir Temperature [150 dec-C]'
-
-        assert hip_ra.rejection_temperature.value == 25.0
-        assert hip_ra.reservoir_porosity.value == 18.0
-        assert hip_ra.reservoir_area.value == 81.0
-        assert hip_ra.reservoir_thickness.value == 0.286
-        assert hip_ra.reservoir_life_cycle.value == 30
-        assert hip_ra.rock_heat_capacity.value == 2840000000000.0
-        assert hip_ra.fluid_heat_capacity.value == -1.0
-        assert hip_ra.fluid_density.value == -1.0
-        assert hip_ra.rock_density.value == 2550000000000.0
-        assert hip_ra.rock_recoverable_heat.value == -1.0
-        assert hip_ra.rejection_temperature.value == 25.0
-
-        # FIXME TODO determine if these are applicable
-        # assert hip_ra.WaterContent.value == 18.0
-        # assert hip_ra.RockContent.value == 82.0
-        # assert hip_ra.rejection_entropy.value == 0.367
-        # assert hip_ra.rejection_enthalpy.value == 104.8
-
-    def test_logger_initialization(self):
-        hip_ra = self._new_hip_ra_test_instance(enable_hip_ra_logging_config=True)
-        assert hip_ra.logger.name == 'root'
-        assert hip_ra.logger.level == logging.INFO
-        assert hip_ra.logger.isEnabledFor(logging.INFO) is True
-
-    def test_handling_input_file_not_found(self):
-        client = HipRaClient()
-
-        with self.assertRaises(RuntimeError) as rex:
-            client.get_hip_ra_result(
-                HipRaInputParameters(
-                    from_file_path=Path(tempfile.gettempdir(), f'a-non-existent-file_{uuid.uuid1()!s}.txt')
-                )
-            )
-
-        self.assertIn('Unable to read input file', str(rex.exception))
-        self.assertIn('.txt not found', str(rex.exception))
-
-    def _new_hip_ra_test_instance(self, enable_hip_ra_logging_config=False, pre_re_stash_runner=None) -> HIP_RA:
-        stash_cwd = Path.cwd()
-        stash_sys_argv = sys.argv
-
-        sys.argv = ['']
-
-        try:
-            hip_ra: HIP_RA = HIP_RA(enable_hip_ra_logging_config=enable_hip_ra_logging_config)
-            if pre_re_stash_runner is not None:
-                pre_re_stash_runner(hip_ra)
-            return hip_ra
-        finally:
-            sys.argv = stash_sys_argv
-            os.chdir(stash_cwd)
